:recycle: create a safe, channel index type
`V474::Channel` is an object that is as efficient as a `size_t` but, to
create it, it must be in a valid range. So if it can be created, it can be
used as a channel index without having to do range testing. Functions which
require a channel now specify `V474::Channel` as the parameter.