:recycle: use the `Memory` class from VXPP v3.0
Convert the driver to use the `Memory` class. This class allows a formal
declaration of the hardware so the registers are accessed properly.

The hardware engineer chose a design in which the channel index skips 16
bytes instead of just 2 (the size of the register). This required a new
register class which formally described this unusual addressing mode.