:recycle: improve the register step logic
This commit uses a more correct way to compute the step size between
registers. Since the template allows one to specify the type, we use
`sizeof` to help compute the step so it gets scaled properly to 16.

It also defines a local type to make the code easier to read/understand.